wildberries_project/requirements.txt
fastapi==0.78.0
sqlalchemy==1.4.32
alembic==1.7.5
pydantic==1.9.0
psycopg2-binary==2.9.3
requests==2.27.1
python-dotenv==0.19.2
uvicorn==0.17.5
